Move interfaces fader/reader to an 'interfaces' package and dir
