+++ internal/services/enttec/dmxusbpro/fader.go
@@ -1,3 +1,4 @@
+// Package dmxusbpro provides services to interact with the Enttec DMX USB Pro Controller
 package dmxusbpro
 
 import (
