+++ internal/services/enttec/dmxusbpro/reader.go
@@ -1,4 +1,4 @@
-// Package reader provides tools to READ from DMX
+// Package dmxusbpro provides services to interact with the Enttec DMX USB Pro Controller
 package dmxusbpro
 
 import (
