+++ internal/services/printer/service.go
@@ -2,17 +2,17 @@
 package printer
 
 import (
-	"github.com/H3rby7/dmx-web-go/internal/services/reader"
+	"github.com/H3rby7/dmx-web-go/internal/services/interfaces"
 	log "github.com/sirupsen/logrus"
 )
 
 // DMXLoggerService reads from DMX and logs the received changesets
 type DMXLoggerService struct {
-	reader reader.DMXReader
+	reader interfaces.DMXReader
 }
 
 // NewDMXLoggerService creates a new [DMXLoggerService] instance with proper defaults
-func NewDMXLoggerService(reader reader.DMXReader) *DMXLoggerService {
+func NewDMXLoggerService(reader interfaces.DMXReader) *DMXLoggerService {
 	log.Debugf("Creating new DMXLoggerService")
 	b := &DMXLoggerService{
 		reader: reader,
